Refactor tests into smaller files and add shared header
